server/requirements/base.txt
Click==7.0
coreapi==2.3.3
Django==2.1.1
django-cors-headers==2.4.0
django-dotenv==1.4.2
django-python3-ldap==0.11.2
djangorestframework==3.8.2
mysqlclient==1.3.14
Pillow==5.2.0
requests==2.19.1
responses==0.9.0
python-dateutil==2.7.5
django-model-utils==3.1.2
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
ics==0.4
